Move network protocol implementation behind an interface
