+++ src/network/CMakeLists.txt
@@ -1,10 +1,11 @@
 set(common_network_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/address.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/connection.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/connectionthreads.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/mtp/impl.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/mtp/threads.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/networkpacket.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/serverpackethandler.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/serveropcodes.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/serverpackethandler.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/socket.cpp
 	PARENT_SCOPE
 )
